+++ changelog.md
@@ -1,5 +1,21 @@
 # Archive Title Changelog
 
+## 1.0.2, 20241003
+
+### Updated
+- Plugin info
+
+### Fixed
+- Removing obsolete `load_plugin_textdomain()`
+
+### File updates
+	archive-title.php
+	changelog.md
+	readme.txt
+	includes/classes/class-archive-title.php
+	languages/*.*
+
+
 ## 1.0.1, 20230524
 
 ### Fixed